tcz: add options so we can fetch all the tcz files
This will allow us to build a NiChrome image ahead of time
that contains all the packages,
i.e. avoid the need to fetch the images after we boot.

Note that fetching after boot will still work. If we
want a package after boot and it's not there, and the
net is up, tcz will fetch it. But if it's there, there's
no need to fetch and tcz won't.

Hence, we don't need to change xinit
one bit, but we can make its life much easier by preloading
a set of tcz's on the usb stick.

The -i switch, which defaults to true, determines whether we
install the packages, i.e. do the loopback mount and create
symlinks.

The -r switch, which defaults to /tcz, lets us determine where the
packages go.

Suppose in NiChrome we have a list of packages we want to
make available on ROOT-A. Suppose it's just bash
and opera-12. To start, in the NiChrome usb program, we do this:

tcz -r `pwd`/tcz -i=false base opera-12

This will give us a tcz in $pwd that has structure we need for
eventual setup in the ramfs, i.e. files like this
tcz/8.x/x86_64/tcz/bash.tcz

The eventual thing I'm going to try is that ROOT-A won't even
be a file system; it will be a cpio archive containing tcz/ and
anything else we need in the ramfs. This is nice since it
means we don't need DM-verity; we can just sign the cpio archive
and test it on boot. There's also no fsck or other issues
to worry about. We'll see how it goes. I'll be testing more and more
of this next week.

But these changes are useful not matter what we do on NiChrome.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>