Exit command when no args passed
I noticed in some commands where you might not expect a blocking
action we were printing the usage but continuing on with execution.

This meant that some of these commands would throw errors in go routines
or try to make directories etc after printing the usage string.

Combed through the commands and tried to exit quick if no arg is passed
and an arg is required. In some cases add exit codes similar to behavior
of existing commands.