xinit: remove ssh startup, speed up x startup
xinit really should not be in the business of starting sshd.
We can define another tool for that later.

We can test for the existence of /tmp/.X*/X? (the X socket)
and once that is there, start X clients. That's much better than
the 5 second timeout.

With these changes startup is much faster. X display is up in 8 seconds
and aterm and opera are there in 10. There's a few more seconds that will
be shaved off once we are on the MMC. I think we can get to 4 or 5.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>