bulk gofmt -s update
This change is 100% gofmt generated.  The gometalinter and goreportcard
highlighted this set of files as "not gofmted with -s (gofmt)", so they're
run through gofmt to correct the noted issues.

Most changes are minimal stylistic changes, and gofmt is _the_ style
authority.

The changes in cmds/stty/termios_linux.go and pkg/gpt/data.go are
perhaps less obvious.  In both cases the map definition has a
detailed explicit type, which makes the stating the type in the
instantiations redudant and thus gofmt removes that leading text.

Signed-off-by: Tim Pepper <tpepper@vmware.com>