Remove the -test switch from ramfs.go and fix README.md
We've come too far down the road of working initramfs for the
-test switch to be meaningful any more. A pity, but the kinds of
things it has to do to make the test useful are just more than
linux can stand.

remove the -test switch, and a bunch of code; fix README.md.

While we're at it, clean up guessgopath to just die if GOPATH is
not set.

I'm cleaning this up because it was causing trouble for new users.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>