lsmod: Add lsmod command
This really is just a clone of lsmod(8).

Limitations: The formatting should be "pixel perfect" to the original,
which it is for the basic format. Used modules, while all of them are
printed correctly, have a different order than lsmod(8). Anybody parsing
the output and relying on a particular order is lost anyways ;-).

Signed-off-by: Roland Kammerer <dev.rck@gmail.com>