cbmem tool
This tool is much like the coreboot cbmem tool, but in my
view better in some ways. The default behavior is to produce
a complete JSON dump of cbmem. It prints more LB table entries than the
standard cbmem tool and the code is a lot simpler.

Creating this tool found a bug in coreboot.

The definition of the constants and structures can be found in
https://github.com/coreboot/coreboot/blob/master/payloads/libpayload/include/cbfs_core.h
and
https://github.com/coreboot/coreboot/blob/master/src/commonlib/include/commonlib/coreboot_tables.h

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>