rm: Fix out of bounds access; accept 'Y'
So far when "rm -i" was called and the user just hit enter, an empty
string was accessed at index 0. If we use ReadString() instead of a
Scanner like "cp", we avoid that because then we get at least the '\n'.
To me that was nicer than to keep the Scanner and check for "" or
len()==0.

While at it, also accept 'Y' as input, which is the same what coreutils
do.

Signed-off-by: Roland Kammerer <dev.rck@gmail.com>