scripts/ramfs: convert ramfs script to use cpio package.
This gets rid of the gross devtmpfs binary blob in scripts/ramfs.go, and
the calls to cpio as a subprocess. Removes the -d and -removedir flags.

The temporary directory is now not used as a staging ground for the
initramfs; so it does not contain the entire initramfs. Not removing the
temp dir only existed as a debugging measure, but that has no point
anymore.

This is the first step in unifying the ramfs and bb scripts.