README: provide a working bb invocation
The existing invocation is missing a glob on the end that causes a build
failure.

Beyond that it is lacking in context of which directory it must be run
from to work correctly, and the required packages for elvish.

I don't know how best to explain the need to go get boltdb/bolt. It
seems to be a bug in the build tool that it doesn't look in the proper
vendor directories.

Updates #420

Signed-off-by: Michael Pratt <michael@pratt.im>