dhclient: honor OptionDomainNameServer;use RouteAdd and RouteReplace
This change makes a dramatic improvement in the
dhcpclient. default routes work correctly now and
nameservers are set up as they should be.

We use RouteAdd and if that fails with EEXIST, we use RouteReplace.

NiChrome is far more usable.

Also, fix a long standing error in the test. It forced an error
and wanted an exit with no error!

Tested via test and on a Chromebook with NiChrome.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>