make it so dynamic and static can live together in NiChrome
Add a new directory, /bbin, and put busybox builds there.
Have init search globs, not paths, so that we can pick up,
e.g., NiChrome commands. Have installcommand use the same
glob for finding command source.

For now, since we hit a bug in the Go toolchain (22819),
re enable printing commands from Go build. This bug is
supposed to be fixed in 1.10.

This has been tested with NiChrome and seems to work ok.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>