cmds/rm: Add -f force flag
The force flags disables interactive mode (if enabled) and ignores
errors when trying to delete nonexistent file(s).

This allows to invoke `rm` on multiple files and be sure that they will
all be deleted even if one or more is missing. Suppose we have a folder
`foo/` with files `a` and `c` in there. On `rm foo/a foo/b foo/c` only
`foo/a` will get deleted, because `foo/b` is missing and thus the
command will fail. On `rm -f foo/a foo/b foo/c` on the other hand, the
command will succeed and both `foo/a` and `foo/c` will be deleted.

Signed-off-by: Kiril Vladimiroff <kiril@vladimiroff.org>