add one more glob to pkg/uroot/util/root.go for NiChrome
This is one more pattern typical of Go packages
and ensures that the symlink for /ubin/wingo gets
created.

We *could* replace all this with a walk, save that
it seems to create more false positives.

Possibly, what it should do is a walk, then an ast, and see
if package main is defined in that directory, then add a symlink?

That's for another day, and another PR, I think.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>