boot: change parsing code
I think this is easier to deal with: the boot entries from the file get stored in a
map. To get the default entry, you can look at map["default"] and to get the nth entry look
at map["2"] or whatever.

Seems to work, including the panic cases.

but see what you think.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>