pci: allow user to specify the glob
This may still fail the convenience test, as one could argue you
must name the glob exactly (e.g. 0000:02:00.0) but otoh you can say
pci -s *02:00*
and it works.
That's not much worse, I guess, than
pci -s 0:2.0? Or is it?

Plus
pci *18*
gets all devices with an 18. So far it's inconvenient in theory,
convenient in practice.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>