pci pkg/cmd: simplify printing and name lookup
To lookup a name, users should not have to be aware of maps.
PCI devices now have a default VendorName/DeviceName that is numeric.
To map them to names, one can call p.SetVendorDeviceName, which
also works on []*PCI, i.e. Devices.

Since no error checking was done on the parse (it should be done at
build time anyway), remove the error return and simplify that logic.

ToString() is no longer needed, so go with String.

This is all in preparation to make it easy to dump config space.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>