pci pkg/cmd: simplify printing and name lookup (#494)
* pci: allow user to specify the glob

This may still fail the convenience test, as one could argue you
must name the glob exactly (e.g. 0000:02:00.0) but otoh you can say
pci -s *02:00*
and it works.
That's not much worse, I guess, than
pci -s 0:2.0? Or is it?

Plus
pci *18*
gets all devices with an 18. So far it's inconvenient in theory,
convenient in practice.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* pci pkg/cmd: simplify printing and name lookup

To lookup a name, users should not have to be aware of maps.
PCI devices now have a default VendorName/DeviceName that is numeric.
To map them to names, one can call p.SetVendorDeviceName, which
also works on []*PCI, i.e. Devices.

Since no error checking was done on the parse (it should be done at
build time anyway), remove the error return and simplify that logic.

ToString() is no longer needed, so go with String.

This is all in preparation to make it easy to dump config space.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* pci pkg/cmd: add support for dumping config space

This is very handy.

Symbolic prints come later.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>