find package and command
We need to yank the stringer out of ls for long mode.
But this is actually useful as is.

Due to the way go argument parsing works, it can't follow the find syntax.
This might argue for a different name: I wondered about finder.

With no redirection, errors appear inline, but you can redirection them
via fd 2.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>