uroot package: add : syntax back for extra files
The : syntax lets you do stuff like this:
-files /path/to/prototype/root:etc

It will source from /path/to/prototype/root/etc but
put it into the cpio archive as etc. It's impossible to
live without, I keep learning.

I used SplitN, not Split, as there are valid paths (maybe?)
with : in them. That said, we may find we want a more robust
way to do this.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>