pkg/uroot: add tests, some documentation.
Remove pkg/ramfs, which was only used by pkg/uroot, and a strange
abstraction.

Factor out code that decides what files make it into the initramfs, and
make archivers be as dumb as possible.

Adds a directory archiver (only usable as root, as some initramfs files
can only be created as root).

Signed-off-by: Christopher Koch <chrisko@google.com>