Fix switch_root
This fixes/changes a number of things in switch_root:
- Changed overall order to what seems to be the universally-agreed-upon way to do switch_root:
  1. Move the special mounts (/proc, /sys, ...)
  2. Keep a file descriptor for /
  3. Move the root mount
  4. Chroot to .
  5. Delete old root with file descriptor
  6. Exec init
- Use real exec (syscall), we were not able to get the NEWPID fork magic
  to work reliably. Also the old exec code used flags that are mutably
  exclusive?!
- Rewrote the deletion code to be based on file descriptors
- Handle symlinks in deletion code (would previously crash)
- Also migrate /run, if found
- Root mount move no longer hardcodes 'ext4' fs type, so should work on
  all root fs now.

Given the fact that mutually exclusive flags were used, I have some
suspicion the code never ran in this state. With this version, we
successfully switch_rooted into a vanilla CentOS, and systemd started up
just fine.

We copy-pasted two snippets from the go-runtime (we put them in utils.go).
One of them is private in the runtime, and the other one is not
available in any stable version yet. Both can obviously be removed as
soon as the runtime equivalents become available.