Add Error Handling for wifi cmd when called with no arguments
Before this commit, when call wifi with no arugments, the goroutine will
panic and give you an error.
Now, when call wifi with no arguments, the flag.Usage() function will get
call and the standard usage message will be printed. This is a more
graceful way of handing this error

Add Test Cases for wifi cmd:
The test case includes:
 * No Flags, No Args
 * Flags,No Args

Change Usage message

Signed-off-by: Thien M. Hoang <thienhoang@google.com>