Add the forth package
This may seem crazy but it's actually a handy little interpretive
language. I had forgotten about this until trammell mentioned forth.

We used this to parse command lines in a 500 node cluster, where
each node got the same command and needed to turn it into a pod
(via / and %) and a node (via other operations).

If this gets in I'll write a trivial forth cli -- or maybe just
make rush into a forth shell?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>