u-root: Add templates
A template is a name that expands to more names.

It is defined in a map in the u-root command as a
map[string][]string

So you might, for example, define
"facebook": "/bin/ed"

you can use more than one template, and they can be freely
mixed with real names:
./u-root facebook cmds/wifi

If there are duplicates that is not a problem.
You can, e.g.
u-root core all cmds/cat

and only one copy of each goes in, of course.

Template names can't reference template names (yet), but we can add this
if people want it; it's not terribly hard.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>