init: add a loop to reap dead child processes
In several years of using u-root nobody noticed this.
Leave it to Trammell to find it out.

I also took the opportunity to tidy things up a bit, and reduce
the nesting in the main for loop.

If we're going to run systemd nothing else matters.
Might as well do that check once, rather than every time
through the loop, since that would lead to odd behavior
should the systemd exec fail.

This is in need of review and testing.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>