xcmds: create xcmds directory, mv ash to it
we have lots of "in progress" commands and packages and it's time
to break things out.

xcmds is where we put experimental stuff there is no clear need for.

I want to go with xcmds/ instead of cmds/x/ because the latter
is going to break every glob everyone is using. xcmds nicely
separates things out in a way that is easy to glob.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>