xcmds: create xcmds directory, mv unused/experimental commands to it
we have lots of "in progress" or "not widely used" or "nice try but" type
commands and packages and it's time
to break things out.

xcmds is where we put experimental stuff there is no clear need for.

I want to go with xcmds/ instead of cmds/x/ because the latter
is going to break every glob everyone is using. xcmds nicely
separates things out in a way that is easy to glob.

We all agreed this was ok at the friday the 13th meeting.
Sorry, ed, you're just not ready.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>