rm: update to use github.com/spf13/pflag
We now get reasonable flag handling for rm without
the hack of having a separate "-rf" flag.

We also get long flag names.

If this seems reasonable we should probably use it for our
other tools. The Go flag package is not compatible with
people's expectations.

Here is the change in sizes with this package in rm, for static and dynamic
(larger is dynamic)

without pflag
-rwxr-xr-x 1 rminnich rminnich 37311832 Jul  6 09:43 /tmp/initramfs.linux_amd64.cpio
-rwxr-xr-x 1 rminnich rminnich 12219892 Jul  6 09:43 /tmp/initramfs.linux_amd64.cpio

with pflag
-rwxr-xr-x 1 rminnich rminnich 37465076 Jul  6 09:45 /tmp/initramfs.linux_amd64.cpio
-rwxr-xr-x 1 rminnich rminnich 12510708 Jul  6 09:46 /tmp/initramfs.linux_amd64.cpio

The ogier/pflag was a 150K growth; spf13 is about 300K, surprisingly, but probably worth it?
Or we can revive ogier and go with the smaller footprint if it gets to be an issue.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>