Move stty support functions to termios
We need them for other things anyway.

This necessitated changes to termios, and pty
code, but it was all broken anyay.

We include the testramfs.go program as an example
of both interactive and scripted usage. You can now
script a ramfs running in a container with
private name spaces.

stty command is now considerably simpler.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>