bzimage command and package support for same
bzImage dump bzImage
bzImage initramfs bzImage initramfs newbzImage

In support of this, add an init.cpio for testing.
While we're at it, add the stuff we use to make it.

In conjunction with the kernel in testdata/bzImage, this creates
a working kernel with an initramfs.

One limit: we require that the kernel have a fixed load address,
so we can set that address in the LinuxHeader (bootparams).
I don't think this is a big deal. KASLR is really not needed
for a firmware kernel.

CONFIG is to create testdata/bzImage.

We also don't support placing the kernel and initramfs
above the 4G limit. To do that we'd need to add code
setting
        ExtRamdiskImage
        ExtRamdiskSize
in the LinuxHeader. That's just crazy.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>