pkg/uroot: add unit tests for CreateInitramfs.
Also changes when files are added to the archive.

Previously, directories were added to a list of files to be added to the
archive, and upon writing, the directories were traversed to find
children to write to the archive.

Now, directories are traversed when added to the archive object, and
writing an archive to a file is a simple writing operation.
files_test.go also reflects this.