gang of four: allow u-root to start with 4 binaries not 5
In our never ending quest to shrink the images, no matter what
it takes, this PR shows how we might do so.

The idea is to drop a file, zzzzinit.go, into go/src/cmd/build,
which is
1) the last init function run, and
2) the init function.

The function looks at os.Args[0] and, if it is /init, proceeds
to build installcommand, then runs /ubin/init, which gets built
and is then run.

So the go command is now init, with special behavior when it is
called as such.

The zzzzinit.go is not placed in the command directory unless
GOROOT is set, and the file does not already exist. It is removed
once u-root is run. If you run two u-root's at the same time one will fail.

Results:
With compiling installcommand on boot, it's dramatically slower, 13 seconds
vs 5 seconds

but it's also 35M vs 37.5M

And note the size of the savings is constant: installcommand is the same size
no matter what tools you include.

So some space constrained environments might find this useful. That said, I'm sure this
one needs some, um, self criticism sessions.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>