prune elvish
This is an attempt to shrink and simplify elvish.
Consider it a checkpoint commit.

It has impact in the sense of cutting down elvish complexity
dramatically, but so far I don't see huge impact on size:
compressed images

-rwxr-xr-x 1 rminnich rminnich 15252220 Oct 6 10:04 /tmp/initramfs.linux_amd64.cpio
-rwxr-xr-x 1 rminnich rminnich 5695727 Oct 6 10:04 /tmp/initramfs.linux_amd64.cpio.gz

-rwxr-xr-x 1 rminnich rminnich 15022780 Oct 6 10:03 /tmp/smallelvish.cpio
-rwxr-xr-x 1 rminnich rminnich 5610432 Oct 6 10:03 /tmp/smallelvish.cpio.gz

so only 200K or so uncompressed and 80k compressed.

OTOH the stripped elvish does shrink from 4.7 to 3.9M so that's something -- almost 20%.

At the same time I may be on the edge of a lot of removal which will shrink it greatly,
but it's hard to tell. There's lots of stuff in elvish that seems to come from a void*
world (e.g. prolific use of interface{}, all the hashmap stuff)
and it would be nice to make that disapperar

But we do remove JSON from elvish, and we do remove all of xiaq's vendored packages,
and we do remove use of unsafe, and those may have value in and of themselves.

I nastily delete or inactivate a bunch of tests I doubt we need. As this work continues
we can figure out what to nuke.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>