redo rsdp and remove unneeded goroutine; add a test
includes work from jean-marie:
Fix rsdp with Go 1.11, and imply early exit of kmesg read if RSDP pointer value
is found. This is a mandatory patch to make work Ubuntu 16.04 installer
booting on Winterfell node, otherwise the installer kernel is unable to properly
detect IRQ mapping

The trick to making this work is to open a unix fd with O_NONBLOCK
and open an os.File with that fd. Then all the messiness
goes away.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>