Add comments for hosted mode; add tests to ensure it keeps working
It is possible to have an initramfs for "hosted" mode, i.e. a
mode in which u-root is installed onto an existing system
such as a chromebook.

In that case we should not create defaultsh and init symlinks,
or try to populate /dev, /usr, /lib, and so on.

This works today with the right commandline usage:
go run u-root.go -base=/dev/null -defaultsh="" -initcmd=""

Add a bit more cmdline documentation for 'hosted' mode,
and add two tests to make sure nobody breaks it by accident
in the future. NiChrome will be using this mode.

Kind of cool: I did not know this would even work but the command
needed no changes.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>