abi: change flag types
Flag is now an interface, with Match(), String(), and Mask()
functions. We provide three implementations:
BitSet, wherein a single bit, if set, matches the value
Value, where the entire uint64 is used for the match
Field, where a subset of bits, one or more, is used to match and
       then prettyprint the value as a hex number.

Each has a different model for clearing bits.
In BitSet, the bit is the mask
In Value, the mask is a constant MaxUint64
In Field, a set of bits (not necessarily continguous) provides the mask

I can now have a FlagSet consisting of, e.g., intermixed BitSet and Field
components. This will suffice to print all the weird cases I can think of.

We may, in future, need more sophisticated field extractors but we don't yet.

This change is sufficient to allow me to parse ChromeOS GPT extra fields.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>