ExtraBuildEnv for qemu and tests
u-bmc has a slightly more complex build system than u-root and needs
to pass some extra environment variables.

For ExtraBuildEnv nothing is using that particular value in u-root, but
as the same Options struct is passed around in u-bmc it would be nice to
add it to that structure.

This does break the abstraction, but as this aligns u-bmc and u-root
tests more visually it's most likely worth it right now.

Signed-off-by: Christian Svensson <bluecmd@google.com>