ps: cleanup, add tests
ps is a mess. It has no tests, it has lots of nasty failures, and it is inflexible.
It also does not take a quick snapshot of the system as it should: some files were
read at much later times than others, and sometimes they vanished.

Further, x printing was disabled at some point for some reason, limiting ps usefulness.

This set of changes
o cleans up some of the big issues with reading files
o adds some tests
o uses glob/ReadFile to get the files and hence removes a lot of readdir operations
  from ps itself (but note, they are still of course being done in glob; there
  is no free lunch since the Unix model is to glob outside the kernel, not in
  the kernel)
o introduces UROOT_PSPATH, which will let one pick a different pspath than /proc
  e.g. you can have
  UROOT_PSPATH=/proc/*/stat:/netproc/*/*/stat
  and it will pick up the proc directories you might have mounted from other hosts
  via sshfs or 9p (I'm going to be using this ...)
  ps will show things like
  1 for local procs
  host/1 for netprocs
o attempts to re-enable the exec test, though it fails with an infinite loop
  what am I doing wrong here? it fork bombs

There is a lot left to do; consider this a downpayment on a substantial
technical debt. It's not done, but it's a start.

ps also had a lot of races since processes come and go.
A non-racy ps is impossible given how linux /proc works but we at least
gather the data in a coherent way now.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>