pkg/uroot: fix the -files rename case
We recently changed lddfiles to return all the files, including
the source. This is arguably far more convenient for lddfiles
package users but caused a problem for uroot.

This change resolves the problem: when we're adding files
from ldd, we don't add the file we ran ldd on, since we
previously added it anyway.

Hence, if we do something like this:
go run u-root.go -build=bb -files /bin/date:d -files /etc/hosts:h

the right set of files is added, as well as the right set of libraries.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>