forth: for v4.0.0, change the api
The fiano futk tool revealed shortcomings in the forth package
we never saw in years of use at Sandia.

Eval will now return an error but leave results on the stack.
EvalPop will return TOS and and an error, and will further
panic if the stack is not empty.

EvalPop is for use in argv() parsing; Eval in programs
and CLIs.

A command is now an array of Cells. For example NewWord
is this:
NewWord(string, Cell, ...Cells)

This just feels a whole lot better, since you can now easily
create new words with strings with embedded white space.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>