Bootconfig fixes/improvements (#79)
* bootconfig: Reboot after loading kernel

PR #65 refactored the bootconfig implementation, and in the process made
it so we only load the kexec kernel, but don't boot it

* bootconfig: Make Boot() fail on reboot failure

If we return from Reboot() without an error, there is something wrong as
it should cause the machine to reboot and thus never return. This logs a
failure and returns an error in such a case

* bootconfig: Catch more cases of missing KexecBin

When kexecbin is nowhere in PATH, the error we obtain from exec is
ErrNotFound, rather than something caught by os.IsNotExist(). Add a
second case to catch those