Add integration test for tcz
The checking for mounts fails, not sure why, it's disabled.
It would almost be nicer if I could suck out all the output from qemu
and then look for the matches. Not sure that's possible but it
would be easier.

Actual mounting deferred at present since the test kernel does not
include loop (!!!!)

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>