ash: incorporate the rush parser
ash is 2.8M without it and 3.0M with it.
Rush is 2.9M
Elvish is 6.1M

initramfs with only elvish and init
-rwxr-xr-x 1 rminnich rminnich 1039444 Dec 18 08:26 /tmp/initramfs.linux_amd64.cpio.xz

initramfs with only ash and init
-rwxr-xr-x 1 rminnich rminnich 550428 Dec 18 08:25 /tmp/initramfs.linux_amd64.cpio.xz

It's not a gigantic saving, but OTOH, the code is a lot more straightforward.
In most cases we'll probably stick with elvish but this is nice to have available if we run
out of space.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>