pox and ldd: resolve issues with ld-so paths
Some libraries do not have an interpreter section
and in that case the ldd package has to guess.
We used to use a basic glob: /lib*/ld-*.so.
Problem is, /lib/ on 64-bit x86_64 systems contains
the 32-bit ld-linux.so, i.e. the default place to look
on 64-bit systems has the 32-bit ld-linux.so. This causes
all kinds of fun the ld-linux.so tries to deal with a
64-bit library. We now check the ELF file type
and fix up the search path accordingly.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>