add plan9 style cpu command
CPU is an ssh client that starts up a shell on a remote machine,
as usual; but, further, makes a namespace of the local machine
available in a private mount rooted at /tmp/cpu on the remote.

On the remote machine, you can access your local files with
cp, moving things back and forth, running binaries, and
so on.

Whereever you go, there your files are.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>