[bb] check if symlink instead of whether resolved filename is "bb"
The logic of resolving symlinks until we encounter "bb" introduced in
aaa8821 also broke a use cases when the
binary generated with makebb is not named "bb" after all. This can be
easily acheved by calling makebb with -o switch.

```
$ go build
$ ./makebb -o test ../../cmds/{ls,df}
$ ln -s makebb ls
lrwxrwxrwx 1 lsiudut lsiudut       4 Jan 30 11:18 ls -> test
-rwxr-xr-x 1 lsiudut lsiudut 8929697 Jan 30 11:14 makebb
-rw-r--r-- 1 lsiudut lsiudut    1112 Jan 30 11:12 makebb.go
-rwxr-xr-x 1 lsiudut lsiudut 2290720 Jan 30 11:18 test
$ ./ls
2019/01/30 11:18:10 os.Args is [test]: you need to specify which command to invoke.
```

I propose to check if resolved path is a symlink and continue traversal if it is. If
not - assume that this is our final bb binary.

Signed-off-by: Lukasz Siudut <lsiudut@gmail.com>