Always add CGO_ENABLED when building the toolchain
This ensures that defaultCGO_ENABLED is set correctly.
This variable is created at build time as part of
building the Go build tool and for whatever
reason we were not getting it set correctly. It makes
sense to hardwire it here since the consequences of it
being wrong are so serious. dynamic mode has been broken
for some time now and we did not know, for example.

This fixes the problems people have had with both testramfs
and qemu using dynamic mode.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>