Avoid error when probing module
The current implementation of Probe tries to load the module first
without checking if it's loaded or if it has dependencies to load first.
In the case of trying to load a module that has dependencies, this
fails by design and the returned error is used to properly load the
module.

This patch introduces parsing /proc/modules to find loaded modules and
removes the logic of special error handling.

Signed-off-by: Nikolas Sepos <nikolas.sepos@gmail.com>