add the ability to do old-school kexec boots without multiboot
This is in pursuit of allowing linuxboot to add an IBFT to an existing
table.

In an idea world, we'd just rewrite the RSDP with a new 64-bit pointer
to the XSDT. BUT:
it's essentially impossible to move the RSDP at present, since the
memory it lives in is protected at the southbridge/ICH/PCH (pick your
generation), so you have to let it point where it points
change the memory it points to. We'll add a new segment which is the
ACPI table with an IBFT added.

To test all this without doing a multiboot kernel, we add a third option
to the kexec command to use the old-school segment-based kexec. It was not
possible to use this until now unless you were doing a multiboot kernel like
esxi, and we don't want to use esxi to debug our ibft experiments.

This way, we can continue to test just be testing Linux and giving it a new
ACPI table to find.

The overall problem (which is actually going to be a mess to fix, involving as it
does 30 years of io chipset hardware) is that in a linuxboot world, where linux
is firmware, a few things have been locked down by the time linux starts, including
the e0000/f0000 areas where rsdp lives. The bits that protect this memory, once
written, can't be cleared. So it acts like RAM, but writes to it are
transparently discarded. LinuxBoot tends to break some rules.

This is all still WIP.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>