ACPI package
This package supports reading and writing of ACPI tables.

There are three models of tables supported:
Raw, tables which just consist of a []byte
Generic, tables which consist of the standard Header and []byte

The third type contains very specific structures; one example
you can see is in sdt.go, where we read in the header, and
also the pointers.

All of the tables implement the Tabler interface, which allows
programs to query values of standard header items such as
signature, length, revision, and so on; as well as
Marshal them out.

It is easy to convert a Raw to a Generic and a Generic to
a more refined table. The latter requires a table-specific
unmarshaler, however.

You can see in types.go that we define a set of standard
scalar types, and use those types in fns.go. For example, there is
a signature type which is just a string. You can use these
types to create table schemas.

I experimented with using structure member tags and this
other model of purpose-made types. The latter proved more
convenient. You can see in ibft.go how they are used.

Since we don't care about most tables, we don't implement
them; we just use Raw.

The ibft.go tables were created directly from the ACPI
spec, via scripts; should we wish to implement more tables,
at some point we should just write a go program to slurp in the
spec and generate structs and constants. It's just too painful
to do it any other way.

One other point. If you want to generate a new table for, e.g.,
kexec, it's pretty certain you'll have to put the [XR]SDT at
the address pointed to by the RSDP, as the RSDP is usually
in write-protected (by the PCH!) RAM. The bits that protect
that RAM are not resettable save by a power-on/reset cycle.
Writes to that area are transparently discarded; see all the
snarky notes in the source about that.
If you are wondering whyMarshalAll is implemented as it is,
now you know: [XR]SDT can't move.

I'm open to suggestions on ways to do this better, moving forward,
but be aware that for now, this works to let us add tables using
our kexec multiboot support in u-root. Be aware that with ACPI,
barely good enough is usually all you can expect. Just see the
ACPICA code some time :-)

This commit also includes a hack to fix the dmesg test because
I can't get it to work any other way.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>